Refactor: Remove unused divide and conquer search module
This commit removes the entire divide and conquer search module as it is no longer in use.
