+++ backend/onyx/agents/agent_search/run_graph.py
@@ -5,14 +5,8 @@ from langchain_core.runnables.schema import CustomStreamEvent
 from langchain_core.runnables.schema import StreamEvent
 from langgraph.graph.state import CompiledStateGraph
 
-from onyx.agents.agent_search.dc_search_analysis.graph_builder import (
-    divide_and_conquer_graph_builder,
-)
-from onyx.agents.agent_search.dc_search_analysis.states import MainInput as DCMainInput
 from onyx.agents.agent_search.dr.graph_builder import dr_graph_builder
 from onyx.agents.agent_search.dr.states import MainInput as DRMainInput
-from onyx.agents.agent_search.kb_search.graph_builder import kb_graph_builder
-from onyx.agents.agent_search.kb_search.states import MainInput as KBMainInput
 from onyx.agents.agent_search.models import GraphConfig
 from onyx.chat.models import AnswerStream
 from onyx.server.query_and_chat.streaming_models import Packet
@@ -21,7 +15,7 @@ from onyx.utils.logger import setup_logger
 
 logger = setup_logger()
 
-GraphInput = DCMainInput | KBMainInput | DRMainInput
+GraphInput = DRMainInput
 
 
 def manage_sync_streaming(
@@ -51,18 +45,6 @@ def run_graph(
         yield cast(Packet, event["data"])
 
 
-def run_kb_graph(
-    config: GraphConfig,
-) -> AnswerStream:
-    graph = kb_graph_builder()
-    compiled_graph = graph.compile()
-    input = KBMainInput(
-        log_messages=[], question=config.inputs.prompt_builder.raw_user_query
-    )
-
-    yield from run_graph(compiled_graph, config, input)
-
-
 def run_dr_graph(
     config: GraphConfig,
 ) -> AnswerStream:
@@ -71,15 +53,3 @@ def run_dr_graph(
     input = DRMainInput(log_messages=[])
 
     yield from run_graph(compiled_graph, config, input)
-
-
-def run_dc_graph(
-    config: GraphConfig,
-) -> AnswerStream:
-    graph = divide_and_conquer_graph_builder()
-    compiled_graph = graph.compile()
-    input = DCMainInput(log_messages=[])
-    config.inputs.prompt_builder.raw_user_query = (
-        config.inputs.prompt_builder.raw_user_query.strip()
-    )
-    return run_graph(compiled_graph, config, input)
